Cleanup broken symlinks that could be leftover
Find and remove broken symlinks in home and config. If things are
removed, there is no good way to know what it is, so cleaning up broken
links is the simple way.
Finding a solution for removing things no longer set is probably needed.
The simple way is not so simple because I like to use dev for other
things that I might not want on multiple machines.

Also store default config location in XDG_CONFIG_HOME if it is empty, so
we can use it for cleaning links later.