Loop for args and run rm in find
Loop for args is much better, doesn't force certain order.
Store args in a variable for use when relaunching due to update.
Add simple help text.
Add error function to take make cleaner error writing.
The -delete flag for the find command is not universally supported.
Change it to an -exec of rm, which is slower and less clean, but since
rm will only be called when something needs deleting, the slowdown
should never be noticed.