Add force link option for locally created files
When modifying a project locally, profile-manager has no way to know
that it was updated, since it compares against the remote. It will never
run the link rules then, so added files would never be linked. Adding a
force option will allow for this to be fixed, albeit manually. An
automatic way would be tricky, and probably isn't worth the edge cases.

Also, don't restart if we are force linking. This does miss a case where
we did actually update ourself when being told to force link, but most
likely that case won't happen, and if it does it won't break anything,
it can just be run again.