Per repo hooks and fix for relaunching
There are now per repo hooks for post_update and post_install. They are
per repo ONLY right now, global should come soon. Simply add a
post_update() and/or post_install() functions to the repo function, and
they will be run after their respective operations. Ideally, only
post_update will be needed, for compiling and stuff, but post_install
could be used to install dependencies, if it is simple enough.

Fix relaunching when updating either profile-manager or its config. This
was a pain. I had forgotten that it is impossible to pass vars back out
of a subshell. Turns out the only reliable way to do it is to pass back
a return value, simple enough.

The tough bug fix was the check itself. Default shell conditionals are
crap, to say the least. Had to go with a full bash way to get it to
behave like I needed.