Fix issues with MacOS
The /bin/bash on MacOS is very old, and this script doesn't support it.
By using env we can load the default Bash which will be a newer version.

Also on MacOS not all files in our home directory can be actually read.
Doesn't make sense to me, as I own them, but rather than try to fix it
just silence the errors.